refactor eject commands
