+++ packages/expo-cli/src/commands/eject/ejectAsync.ts
@@ -1,8 +1,22 @@
+import { getConfig } from '@expo/config';
+import { Versions } from 'xdl';
+
+import CommandError from '../../CommandError';
+import Log from '../../log';
+import { confirmAsync } from '../../utils/prompts';
 import maybeBailOnGitStatusAsync from '../utils/maybeBailOnGitStatusAsync';
 import { promptToClearMalformedNativeProjectsAsync } from './clearNativeFolder';
 import { logNextSteps } from './logNextSteps';
-import { assertPlatforms } from './platformOptions';
-import { EjectAsyncOptions, prebuildAsync } from './prebuildAsync';
+import { assertPlatforms, platformsFromPlatform } from './platformOptions';
+import { EjectAsyncOptions, prebuildAsync } from './prebuildAppAsync';
+
+async function userWantsToEjectWithoutUpgradingAsync() {
+  const answer = await confirmAsync({
+    message: `We recommend upgrading to the latest SDK version before ejecting. SDK 37 introduces support for OTA updates and notifications in ejected projects, and includes many features that make ejecting your project easier. Would you like to continue ejecting anyways?`,
+  });
+
+  return answer;
+}
 
 /**
  * Entry point into the eject process, delegates to other helpers to perform various steps.
@@ -11,7 +25,7 @@ import { EjectAsyncOptions, prebuildAsync } from './prebuildAsync';
  * 2. Prebuild the project
  * 3. Log project info
  */
-export async function ejectAsync(
+async function ejectAsync(
   projectRoot: string,
   { platforms, ...options }: EjectAsyncOptions
 ): Promise<void> {
@@ -22,4 +36,36 @@ export async function ejectAsync(
 
   const results = await prebuildAsync(projectRoot, { platforms, ...options });
   logNextSteps(results);
+}
+
+export async function actionAsync(
+  projectRoot: string,
+  {
+    platform,
+    ...options
+  }: Omit<EjectAsyncOptions, 'platforms'> & {
+    npm?: boolean;
+    platform?: string;
+  }
+) {
+  const { exp } = getConfig(projectRoot);
+
+  if (options.npm) {
+    options.packageManager = 'npm';
+  }
+
+  // Set EXPO_VIEW_DIR to universe/exponent to pull expo view code locally instead of from S3 for ExpoKit
+  if (Versions.lteSdkVersion(exp, '36.0.0')) {
+    if (options.force || (await userWantsToEjectWithoutUpgradingAsync())) {
+      throw new CommandError(
+        `Ejecting to ExpoKit is now deprecated. Upgrade to Expo SDK +37 or downgrade to expo-cli@4.1.3`
+      );
+    }
+  } else {
+    Log.debug('Eject Mode: Latest');
+    await ejectAsync(projectRoot, {
+      ...options,
+      platforms: platformsFromPlatform(platform),
+    } as EjectAsyncOptions);
+  }
 }
