+++ packages/expo-cli/src/commands/index.ts
@@ -8,8 +8,9 @@ const COMMANDS = [
   require('./build'),
   require('./client'),
   require('./credentials'),
-  require('./customize'),
-  require('./eject'),
+  require('./eject/customize'),
+  require('./eject/eject'),
+  require('./eject/prebuild'),
   require('./expokit/bundle-assets'),
   require('./expokit/prepare-detached-build'),
   require('./export'),
@@ -20,7 +21,6 @@ const COMMANDS = [
   require('./info/upgrade'),
   require('./init'),
   require('./install'),
-  require('./prebuild'),
   require('./publish'),
   require('./push'),
   require('./run'),
