+++ packages/expo-cli/src/commands/run/android/runAndroid.ts
@@ -10,7 +10,7 @@ import StatusEventEmitter from '../../../analytics/StatusEventEmitter';
 import getDevClientProperties from '../../../analytics/getDevClientProperties';
 import Log from '../../../log';
 import { promptToClearMalformedNativeProjectsAsync } from '../../eject/clearNativeFolder';
-import { prebuildAsync } from '../../eject/prebuildAsync';
+import { prebuildAsync } from '../../eject/prebuildAppAsync';
 import { installCustomExitHook } from '../../start/installExitHooks';
 import { profileMethod } from '../../utils/profileMethod';
 import { setGlobalDevClientSettingsAsync, startBundlerAsync } from '../ios/startBundlerAsync';
