+++ packages/expo-cli/src/commands/run/ios/runIos.ts
@@ -9,7 +9,7 @@ import StatusEventEmitter from '../../../analytics/StatusEventEmitter';
 import getDevClientProperties from '../../../analytics/getDevClientProperties';
 import Log from '../../../log';
 import { promptToClearMalformedNativeProjectsAsync } from '../../eject/clearNativeFolder';
-import { EjectAsyncOptions, prebuildAsync } from '../../eject/prebuildAsync';
+import { EjectAsyncOptions, prebuildAsync } from '../../eject/prebuildAppAsync';
 import { installCustomExitHook } from '../../start/installExitHooks';
 import { profileMethod } from '../../utils/profileMethod';
 import { setGlobalDevClientSettingsAsync, startBundlerAsync } from '../ios/startBundlerAsync';
